backend/requirements.txt
aiodns==3.0.0
asgiref==3.4.1
astroid==2.6.2
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.4.0
autopep8==1.5.7
beautifulsoup4==4.9.3
boto3==1.18.28
botocore==1.21.28
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-post-office==3.5.3
django-rest-framework==0.1.0
django-ses==2.2.1
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
dnspython==2.1.0
filelock==3.0.12
future==0.18.2
idna==3.2
iniconfig==1.1.1
ipaddress==1.0.23
isort==5.9.2
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.6.0
lxml==4.6.3
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
packaging==21.0
pep8==1.7.1
pluggy==0.13.1
py==1.10.0
py3-validate-email==1.0.1
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.9.3
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.8
wrapt==1.12.1
yarl==1.6.3
celery==4.3.0
flower==0.9.3
redis==3.5.3
django-redis==5.0.0
